creates controllers folder for route controllers
